+++ src/MatComputation.cpp
@@ -6,7 +6,7 @@ normalize the 64-bit double to 32-bit float matrix, to z-score the numbers
 input: starting time point, end time point, number of voxels, number of total time points, the raw matrix, normalized matrix
 output: write results to the normalized matrix, return the length of time points to be computed
 ******************************/
-int getBuf(int start_col, int end_col, int row, int col, double* mat, float* buf)
+int getBuf(int start_col, int end_col, int row, int col, float* mat, float* buf)
 {
 	int i, j, delta_col=0;
 	for (i=0; i<row; i++)
@@ -15,8 +15,8 @@ int getBuf(int start_col, int end_col, int row, int col, double* mat, float* buf
 		double mean=0, sd=0;  // float here is not precise enough to handle
 		for (j=start_col; j<=end_col; j++)
 		{
-			mean += mat[i*col+j];
-			sd += mat[i*col+j] * mat[i*col+j]; // convert to double to avoid overflow
+			mean += (double)mat[i*col+j];
+			sd += (double)mat[i*col+j] * mat[i*col+j]; // convert to double to avoid overflow
       delta_col++;
     }
 		mean /= delta_col;
@@ -48,8 +48,8 @@ CorrMatrix* CorrMatrixComputation(Trial trial, int sr, int step, RawMatrix** mat
   int row1 = matrices1[sid]->row;
   int row2 = matrices2[sid]->row;
   int col = matrices1[sid]->col;  // the column of 1 and 2 should be the same, i.e. the number of TRs of a block
-  double* mat1 = matrices1[sid]->matrix;
-  double* mat2 = matrices2[sid]->matrix;
+  float* mat1 = matrices1[sid]->matrix;
+  float* mat2 = matrices2[sid]->matrix;
   float* buf1 = new float[row1*col]; // col is more than what really need, just in case
   float* buf2 = new float[row2*col]; // col is more than what really need, just in case
   int ml1 = getBuf(sc, ec, row1, col, mat1, buf1);  // get the normalized matrix, return the length of time points to be computed
