+++ src/MatComputation.h
@@ -7,6 +7,6 @@ extern "C" {
 }
 #endif
 
-int getBuf(int start_col, int end_col, int row, int col, double* mat, float* buf);
+int getBuf(int start_col, int end_col, int row, int col, float* mat, float* buf);
 CorrMatrix* CorrMatrixComputation(Trial trial, int sr, int step, RawMatrix** matrices1, RawMatrix** matrices2);
 CorrMatrix** ComputeAllTrialsCorrMatrices(Trial* trials, int nTrials, int sr, int step, RawMatrix** matrices1, RawMatrix** matrices2);
